arm64: DT: dm-verity bootargs for System-as-Root
Android 10 demands System-as-Root, which in turn demands dm-verity.
Our bootloader does not pass anything so we have to define everything
about dm-verity manually.

Note that androidboot.verifiedbootstate has to be orange, otherwise
the verity will always be enforced - that's not one would expect on
an unlocked device, and could cause a lot of inconvenience.

Change-Id: I4292a8d580c65062d124fff7d1ee0649b18bec8b