[HACK] firmware: bundle nouveau firmwares with kernel
This firmwares are needed at a very early stage, before Android init
starts. If these images could not be loaded at that time, nouveau could
panic when it tries to initialize graphics card.

Maybe we can modify nouveau to make it load firmwares later, but this is
the fastest way to get System-as-Root to boot.